+++ framework-rest-resources/framework-stream-rest-resources/src/main/java/uk/gov/justice/services/resources/repository/StreamQueryException.java
@@ -1,0 +1,12 @@
+package uk.gov.justice.services.resources.repository;
+
+public class StreamQueryException extends RuntimeException {
+
+    public StreamQueryException(final String message) {
+        super(message);
+    }
+
+    public StreamQueryException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}