+++ framework-rest-resources/framework-stream-rest-resources/src/main/java/uk/gov/justice/services/resources/rest/model/ErrorResponse.java
@@ -1,0 +1,4 @@
+package uk.gov.justice.services.resources.rest.model;
+
+public record ErrorResponse(String errorMessage) {
+}